revert(experimental): geri dondurulmus gFlashLightBlinkCounter marquee
- Durum makinesi (state machine) yaklasimi donanim dongusuyle uyusmadigi icin kaldirildi.
- Orijinal gFlashLightBlinkCounter + modulo offset mantigi geri getirildi.
- Ghosting/onerme sorununu cozmek icin her karede sol panel memset ile sifirlaniyor.
- Kullanilmayan scheduler.h eklentisi geri alindi.

+++ ui/menu.c
@@ -548,13 +548,18 @@ static void UI_PrintMenuLabelSmallClip(const char *pString, uint8_t Line, uint8_
     UI_PrintStringSmallNormal(buf, 0, 0, Line);
 }
 
-/* HTML-marquee-style continuous left scroll.  The text moves steadily left,
- * disappears, then re-enters from the right after a small gap. */
+/* Continuous left-scroll marquee for long menu labels.
+ *
+ * The text loops smoothly: it shifts left, leaves the panel, then re-enters
+ * from the right after a short blank gap.  Timing uses the existing radio
+ * loop counter gFlashLightBlinkCounter (incremented every 10 ms).  The left
+ * panel rows are cleared before every draw so that character-gap pixels from
+ * the previous frame cannot ghost or overlap. */
 static void UI_PrintMenuLabelScroll(const char *pString, uint8_t Line, uint8_t PanelWidth)
 {
     const uint8_t char_pitch = 8;
     const size_t  len        = strlen(pString);
-    const unsigned int total_width = len * char_pitch;
+    const unsigned int total_width = (unsigned int)(len * char_pitch);
 
     if (total_width <= PanelWidth)
     {
@@ -562,19 +567,19 @@ static void UI_PrintMenuLabelScroll(const char *pString, uint8_t Line, uint8_t P
         return;
     }
 
-    const int gap = 16;                           // blank gap before re-entry
-    const int max_offset = total_width + gap;     // loop length
-    const uint16_t step_period = 15;              // one pixel step every ~150 ms
-    const int offset = (int)((gFlashLightBlinkCounter / step_period) % max_offset);
+    const int      gap        = 16;                         // blank gap before re-entry
+    const int      max_offset = (int)(total_width + gap);   // loop length in pixels
+    const uint16_t step_period = 15;                        // one pixel step every ~150 ms
+    const int      offset     = (int)((gFlashLightBlinkCounter / step_period) % max_offset);
 
-    // Clear the left panel area for this label first, otherwise the 1-pixel
-    // character gaps and the re-entry gap leave old framebuffer pixels behind
-    // and the scrolling text ghosts/overlaps with previous frames.
+    // Clear the left panel area for this label first.  The 1-pixel character
+    // gaps and the re-entry gap would otherwise leave old framebuffer pixels
+    // behind, causing the scrolling text to ghost or overlap with itself.
     memset(gFrameBuffer[Line + 0], 0, PanelWidth);
     memset(gFrameBuffer[Line + 1], 0, PanelWidth);
 
     UI_PrintMenuLabelAt(pString, Line, PanelWidth, offset);
-    UI_PrintMenuLabelAt(pString, Line, PanelWidth, offset - total_width - gap);
+    UI_PrintMenuLabelAt(pString, Line, PanelWidth, offset - (int)total_width - gap);
 }
 
 void UI_DisplayMenu(void)
